refactor: split npm collector as separate module
