+++ src/collector.ts
@@ -9,8 +9,6 @@ import { IPosition, IKeyValueEntry, KeyValueEntry, Variant, ValueType, IDependen
 import { stream_from_string, getGoLangImportsCmd } from './utils';
 import { config } from './config';
 import { exec } from 'child_process';
-import { parse, DocumentCstNode } from "@xml-tools/parser";
-import { buildAst, accept, XMLElement, XMLDocument } from "@xml-tools/ast";
 
 /* Please note :: There was issue with semverRegex usage in the code. During run time, it extracts 
  * version with 'v' prefix, but this is not be behavior of semver in CLI and test environment. 
@@ -212,21 +210,4 @@ class GomodDependencyCollector implements IDependencyCollector {
 
 }
 
-class PackageJsonCollector implements IDependencyCollector {
-    constructor(public classes: Array<string> = ["dependencies"]) {}
-
-    async collect(contents: string): Promise<Array<IDependency>> {
-      const ast = jsonAst(contents);
-      return ast.children.
-              filter(c => this.classes.includes(c.key.value)).
-              flatMap(c => c.value.children).
-              map(c => {
-                  let entry: IKeyValueEntry = new KeyValueEntry(c.key.value, {line: c.key.loc.start.line, column: c.key.loc.start.column + 1});
-                  entry.value = new Variant(ValueType.String, c.value.value);
-                  entry.value_position = {line: c.value.loc.start.line, column: c.value.loc.start.column + 1};
-                  return new Dependency(entry);
-              });
-    }
-}
-
-export { IDependencyCollector, PackageJsonCollector, ReqDependencyCollector, GomodDependencyCollector, IPositionedString, IDependency };
+export { ReqDependencyCollector, GomodDependencyCollector };
