+++ src/server.ts
@@ -8,10 +8,12 @@ import * as fs from 'fs';
 import {
 	IPCMessageReader, IPCMessageWriter, createConnection, IConnection,
 	TextDocuments, InitializeResult, CodeLens, CodeAction, CodeActionKind} from 'vscode-languageserver';
-import { IDependencyCollector, PackageJsonCollector, ReqDependencyCollector, GomodDependencyCollector } from './collector';
+import { ReqDependencyCollector, GomodDependencyCollector } from './collector';
 import { PomXmlDependencyCollector } from './maven.collector';
+import { PackageJsonCollector } from './npm.collector';
 import { SecurityEngine, DiagnosticsPipeline, codeActionsMap } from './consumers';
 import { NoopVulnerabilityAggregator, GolangVulnerabilityAggregator } from './aggregators';
+import { IDependencyCollector } from './types';
 import { AnalyticsSource } from './vulnerability';
 import { config } from './config';
 import fetch from 'node-fetch';
